pyproject.toml
[tool.poetry]
name = "dbt-decodable"
version = "1.3.2"
description = "The Decodable adapter plugin for DBT"
license = "Apache2.0"
authors = [
    "Decodable <info@decodable.co>",
    "Getindata <hello@getindata.com>"
]
readme = "README.md"
repository = "https://github.com/decodableco/dbt-decodable"
keywords = ["decodable", "dbt"]
packages = [
    { include = "dbt" },
    { include = "decodable" },
    { include = "tests" },
]

[tool.poetry.dependencies]
python = ">=3.7.2"
dbt-core = "~1.3.1"

[tool.poetry.dev-dependencies]
black = "~22.3.0"
bumpversion = "*"
dbt-tests-adapter = "^1.3.2"
flake8 = "*"
flaky = "*"
freezegun = "~0.3.12"
ipdb = "*"
pip-tools = "*"
pre-commit = "*"
pyright = ">=1.1"
pytest = ">=7.0"
pytest-dotenv = "*"
pytest-logbook = "*"
pytest-csv = "*"
pytest-xdist = "*"
pytz = "*"
tox = ">=3.13"
twine = "*"
wheel = "*"
pyyaml = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = ["./dbt", "./decodable", "./tests"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    ".venv",
    "*/.venv",
    "build",
    "./typings"
]
stubPath = "./typings"
venv = ".venv"
venvPath = "./"
typeCheckingMode = "strict"
reportMissingTypeStubs = true
reportPrivateImportUsage = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownLambdaType = false
strictListInference = true
strictDictionaryInference = true
strictParameterNoneValue = true
strictSetInference = true
